Validate app version on startup
When a new app version is deployed and the user starts the application,
the browser-side resources (HTML, JS, CSS) are usually updated
automatically. The browser checks if index.html has changed by sending
the Etag of the cached resource. Because the file's content has changed,
the Etag does not match anymore and a new version of the file (where
JS/CSS files with different fingerprints are referenced) is retrieved.

However, there are rare cases where the browser does NOT check for
changes. In FF this may happen when "remember session" is active and the
user exits the browser. In Chrome this may happen when the "auto discard
inactive tabs" feature discards the applications tab. As a consequence,
the application starts up normally, but the browser-side JS/CSS files
may not be up-to-date.

To fix this problem, the server can send the "scout.application.version"
value to the UI with the <scout:version> tag in index.html. The UI sends
it back to the server with the startup request. If the server detects a
mismatch, it returns a special error code (40). This causes the UI to
reload itself, hopefully fixing the issue.

The version is stored on the "scout.app" object. It may also be passed
as an init parameter to the App.init() call.

205169

Change-Id: I68175fd4593bd99b0ebd13fd82e43c53cdbfcb80
Reviewed-on: https://git.eclipse.org/r/91371
Tested-by: Hudson CI
Reviewed-by: Andre Wegmueller <awe@bsiag.com>