Platform: Include invalid property names in exception message
In some runtime environments (i.e. during CI tests), the logger might
not be available when the config.properties file is validated. The names
of the invalid properties would then be unknown which makes the error
hard to fix. Including the invalid property names in the exception
message can help in this case.

Change-Id: I442211b2b404477a524614d54e96c580f9225bd3