Store JSONObject instead of JsonResponse in response history
The response history keeps track of responses never arrived at the
browser and delivers them again if requested. So far, the history
contained JsonResponse objects which can access the model and services
when they are rendered to JSON. The 'sync' request however is neither
executed within a model job nor within a client session.

This change stores response histories as preformatted JSONObjects.

205596

Change-Id: Ia710403e270f7ec5a60f666a71804d4c2ba97c88
Reviewed-on: https://git.eclipse.org/r/91485
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>