Fix tooltip position in popup windows / simplify "detach root" function
Tooltip:
Use correct entrypoint to display tooltip. This fixes a problem where a
tooltip was shown in the main window instead of inside a popup window.

jquery-scout:
Function attachParent() was renamed to the more clear name
"getDetachRoot()" and the implementation was simplified.

Tooltip, BusyIndicator, Popup:
Use "Widget.entryPoint()" instead of statically referring to the
session's $entryPoint property.

204182 (28.)

Change-Id: I63bb2c0d1eacfc03a05171f89345cfa4221d940f