Fixed failing selenium test
Using sendCancelProposal had the side effect that the cancel request
sometimes contains a new proposal chooser. When a user has an invalid
smart field, with an open proposal chooser and then clicks on a menu,
this has the effect, that the menu is closed immediately, because a new
proposal chooser must be displayed. That's why we replaced
cancelProposal with acceptProposal(true) - basically this does the same
as when the user clicks somewhere on the form, outside of the proposal
chooser.