Remove non-displayable menus from events when writing JSON string
Commit fe1fe332a5cad931832193b8e857c9b8ec59a025 fixed an issue when form
fields or menus were set to visibleGranted=false after the JsonAdapter
was attached. This fixed some UI errors. However, there were still some
issues when the now-disposed adapter was referred in an event targeted
for a different (still existing) adapter.

This change puts a special object (FilteredJsonAdapterIds) in the
property change event. The filter is not applied until toJson() is
called on that object.

Example: Menus are set to a group box. The JsonGroupBox generates a
property change event with a list of adapterIds for the corresponding
menus. When the menu is set to visibleGranted=false, but not yet sent to
the UI, the new FilteredJsonAdapterIds removes the adapter ID from the
list when generating the JSON string.

204106

Change-Id: I330d96c47b4be7f2a1408dae54bc120ff91e8f7c
Reviewed-on: https://git.eclipse.org/r/91564
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>