Fixed handling of popup remove (again)
This time the impl. destroyed proper behavior of the multi line
smartfield. Now the special handling required for smart fields in
editable tables is only executed when we're in MODE_CELLEDITOR. 

Also the logic was changed form 'accept proposal' to 'cancel proposal'
because otherwise acceptProposal is called twice, when the popup is
closed normally.