Prevent pointer events on iframes while dragging
When dragging a dialog over an iframe, the iframe might get the
mousemove event and the dragging stops. To prevent this behavior,
$.fn.makeDraggable() adds a CSS class to all iframes when the mouse
button is pressed. When the mouse button is released, the class is
removed again. The class temporarily sets "pointer-events" to "none"
causing the iframe not to receive (and swallow) any mouse events while
dragging.

- SplitBox
- Splitter

[CP from 6.1.x]

Change-Id: Ief77cc2f9b4508deff3493e26b5c37a6f4be1378