Editable table: Don't try to parse the value in case of an error
Use case: 
Delete the text in a cell of a date column. Enter an invalid date. The
text will be the invalid date, the value null and an error status will
be set and shown to the user. In that case the value must not be parsed
because it would fail and an exception would be thrown.

205693