WizardProgressField: Fix background color of last step
I some cases, the background-color of the last step was not updated
correctly. The problem was, when the activeStepIndex was changed before
the step itself was updated (e.g. new text). Instead of only updating
it, when the activeStepIndex changes, the color is now always updated if
the step classes are changed.

Change-Id: I9b8ddf7df8f7f68b2d94754bb34d788a29654560