Tree: fixed another insert position bug in tree.
This time a bug occurred when a childNodeOrderChanged has been received
and the affected node was an expanded subtree. The solution is to check
whether or not the parentNode of a node is already contained in the
insert batch. If that's the case, the insert position of the whole batch
should not be changed.

Cherry picked from releases/6.1.x commit
32360ae7b4ec63cf987f688696d870bf42fbfa5b