iOS: Downloads don't work in homescreen mode
Opening the downloads in new window does not work in
homescreen/standalone mode because the window will be opened in safari
which creates a new http session. Because the downloads are linked to
the http session they cannot be downloaded using safari -> Open in the
same window. 
This works but is not ideal either because there is no back button when
the download is shown. The user has to go to the home screen and open
the app again.

205734