iOS: Use persistent cookie and client session in home screen mode
If the app is running in iOS home screen mode the HTTP session will
be lost whenever the user leaves the app (e.g. switches to another app
or just downloads a file). This means he has to login again and navigate
to the previous location, again. 
To avoid this a persistent mode has been introduced. This means if the
app is running in the home screen mode, a persistent session cookie is
created so that the same http session may be used the next time the app
is activated. Also, the client session id will be put in the local
storage instead of the session storage. This makes sure the same client
session as before is used. Deep link handling is deactivated in that
case otherwise it would always navigate to the url which was active when
"add to home screen" was pressed.

205734