Change TODOs in jquery-2.1.4.js to TO.DO
jquery-2.1.4.js is an external library. The contained "TODO" markers are
part of the officially released version. We do not want them to show up
in our own task lists, therefore they were renamed to "TO.DO".

The same was previously done with jquery-ui-1.11.2.js. A similar "BSI
changes" comment was added on jquery-2.1.4.js.

Change-Id: I63f513a52b4576550ccc5f5d01e3e10381f45188