Fixed JS error when cell editor was open and user clicked on navigation
The error occurred when the cell editor popup was open and the user
clicked on an item in the outline navigation and then returned to the
form with the table. In that case the form is removed and rendered
again. We fixed almost the same bug before, but for detach/attach. We
simply remove the cell editor on the model when remove or detach
happens.

See ticket #205994