Bugfix for issues occurred after connection has been lost.
The issue occurred when connection has been lost while a user clicked on
a smart field or on an editable cell and then navigated to another form
or view. When the connection has been re-established and the initial
form was not rendered or attached anymore, that form tried to open a
popup (for proposal chooser or cell editor) which failed because the
widget was in a detached state - so some DOM operations like accessing
the active document or trying to figure out the position of an element
fail.

The fix checks with the new function isAttachedAndRendered() if the
widget and all of its parents are still valid and open the popup only if
the widget is valid (=attached and rendered).

See ticket #206049