Added request queue and sequence no. to detect duplicate requests.
This should fix some problems where the client has sent a request #n to
the server, the server processes that request, but then the response is
not returned to the client because the client goes offline. When the
client goes online again, it tries to send request #n again, but since
the server has already processed that request, processing the request
again, can lead to various errors (e.g. adapter is already disposed).

With the new logic, each request has a unique sequence no. (similar to
the existing response sequence no.) which allows us to detect and
prevent repeated execution on the server.

205842

Change-Id: Idcb2cc4d3d2d86ba86f4d60e9350aae6026a2e34
Reviewed-on: https://git.eclipse.org/r/92610
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>