Make LookupCalls support "MultilineText"
This change introduces the ability to define the already known concept
of "MultilineText"-ContentAssisFields to a LookapCall itself. This gives
the possibility to display a "MultilineText"-LookupRow in for example a
non "MultilineText"-SmartField.

Technically the ContentAssisFields JS implementation splits the revieced
LookupRows on the first new line to hold the first line in the field
itself and store all additional lines in a variable. The
"MultilineText"-Version renders this additional variable to the visible
multiline area. The normal version just keeps the textual data to be
able to fulfill _readDisplayText() later.

Note: It is not possible to set the LookupCall of a
"MultilineText"-ContentAssisFields to non-multiline.

179688

Change-Id: I96a093cbd205562e35b8df3e1725e279d849e39d
Signed-off-by: Michael Schmuki <michael.schmuki@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/92593
Tested-by: Hudson CI
Reviewed-by: Andre Wegmueller <awe@bsiag.com>