FormField: tooltip missing when there is a menu on the field
The status of a form field either shows the tooltip or a context menu.
If the field has both, a tooltip and menus, only the menus were shown.

The tooltip is now able to show menus as well. So in the above case a
tooltip will be opened and the menus rendered into the tooltip.

Also added a new icon to indicate that the form field has a tooltip and
menus.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=484753