Improved position and max-size of message box.
When the pref. size of a message box (MB) is larger then the screen
size, the size of the MB is adjusted, so it fits into the screen. Also
the content of the MB will be scrollable.

Added static functions on DialogLayout for shared code between MBs and
existing dialogs.

Also added CSS min-width to MB, because MBs with few content didn't look
good.

162168