WrappedFormField: Allow closing/starting inner form
If the inner form is not managed by AbstractWrappedFormField, the form
can be closed and restarted again. Because closing the form
automatically disposes the JSON adapter, a new adapter is created when
the form is started again. The PROP_INNER_FORM property does not change
its value, but the JSON adapter must behave as if the inner form was set
to null.

144072

Change-Id: I45e7c5d841552873ffc618ef05082a32a54e825c
Signed-off-by: Otmar Caduff <otmar.caduff@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/93125
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>