FormField: tooltip and context menu must not be shown simultaneously
If the tooltip contains menus, an already rendered exception occurs if
the context menu should be shown at the same time. This may happen if
the user enters an invalid text and clicks on the status icon -> The
context menu is shown due to the click and the tooltip due to the
invalid value -> prevent that.