Table: allow cell text even if cell value is null
Use case: A column wants to display '-empty-' if there is no cell value.
This was possible in 6.0 but is not possible in 6.1 anymore. 

The UI needs a value for every cell but does not know whether the server
does not send the value because it is null, because it is an
optimization (server does not send the value if it is equals to the
text) or because the column cannot send a value (which is the case for
custom columns).

The UI server now sends value: null in such cases compared to not
sending the property at all.