Fix and harmonize mousedown handling for form fields
- Some fields (smartfield) used click and some fields (datefield) mouse
down to open the popup -> now all fields use mouse down
- Same applies for the field icon clicks / mousedown.
- With commit 771cca5ed3dcbe6d3fdfdb76ef599af389126eb9 a focus issue was
solved, but unfortunately not for IE. In IE, the focus manager tried to
focus another element than the field because _handleIEEvent was
executed. To avoid this, 'unfocusable' css class is used instead of
event.preventDefault().
- Also fixed issue that the field was not focused when it is opened in
touch popup and the icon is clicked.

174228