Fix UI error when closing wizard container form ("Could not resolve event target")
The error was caused by a "Form Closed" event that arrived at (and was
handled by) an already disposed JsonWrappedFormField. It was introduced
with commit e354238488bd6edd2d34bd522883e9af198ac845.

The listener to the inner form must be correctly registered and
unregistered at attachModel() and detachModel() callbacks.

[CP from 6.1.x]

Change-Id: Ibe0a26d95d8c8443767eede8c0b2c1311b5dc5f3