IE: Prevent Error when holding [CTRL]+[C] for more than 0.5s
Because windows forbids concurrent access to the clipboard, a possible
exception is thrown on 'myWindow.clipboardData' in IE (see Remarks on
https://msdn.microsoft.com/en-us/library/windows/desktop/ms649048(v=vs.85).aspx).
Because of this behavior a failed clipboard access will just be logged
but not presented to the user.

Also I noticed that selectAllText() on the clipboard field has no effect
in IE. This change fixes this by using the selection API inside an
setTimeout-callback.

Change-Id: I08037b9acc0dbdfb97f2012a4101609ef3d8bb7d
Signed-off-by: Michael Schmuki <michael.schmuki@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/93508
Tested-by: Hudson CI
Reviewed-by: Matthias Otterbach <matthias.otterbach@bsi-software.com>