Fix coalesce functions (use "target" instead of "id")
Event coalesce functions should check if the previous event's target is
the same adapter as the current event's target. The target is specified
in the "target" property, not in the "id" property.

Added adapter check in DesktopAdapter.js and SmartField.js.

Added coalesce function in ValueFieldAdapter.js to coalesce
'displayTextChanged' events.

Change-Id: I073784cc4fb9a3b46a4a9653dccb75ecbd13c978