Fix size and alignment of context menu popups with PNG icons
Because the size of images is not known until the image has been loaded,
the context menu popup has to be relayouted when the 'load' (or 'error')
event occurs.

Note: This is a temporary solution for context menu popups only. A
better fix would be that _all_ <img> elements invalidate the layout tree
after the 'load' or 'error' event. However, this cannot be implemented
at the moment, because the menubar's layout always recreates the <img>
tags while layouting, thus retriggering the 'load' events (= endless
loop). We intend to find a general solution in a future Scout version,
a corresponding TODO marker was added to the code.

180815

Change-Id: I3e7e84f3a239e5cd71ccdc5f0b888cd1ebe6ec9b