Fixed problem with wrong coloring (child-of-selected) in tree.
addClass('child-of-selected') was done in various places in the code.
This seems to be wrong. Refactored code so we preserve the
child-of-selected class, like parent-of-selected.

Additionally moved _updateItemPath down in Tree.js and execute it after
filters for nodes have been updated. Because node.isFilterAccepted()
(see addToGroup) is in the wrong state otherwise node._decorate is not
called.