Add (currently ignored) test for rare problem in TreeEventBuffer
In certain situations, the TreeEventBuffer fails to remove some events
for deleted nodes, which causes exceptions in JsonTree ("Unknown node").

The situation is as follows:
- Insert a node "A" in a a tree, then a second new node "B" as child
  node of the first new node. The TreeEventBuffer will compact those
  two events to one single NODES_INSERTED event.
- In the same request, remove all child nodes of "A", then all child
  nodes of "A"'s parent. The event buffer now contains three events.
If the events are coalesced, the first ALL_CHILD_NODES_DELETED event
will remain in the buffer, because the links between A and B are already
deleted (and the "real" nodes are referenced in the events).

There is currently no solution to the problem, therefore the unit tests
that can reproduce the problem are currently @Ignored.

Change-Id: I6d31ea5e2ee90a93f23f711c885ad28feab71e05