DateFormat: Pass accepted term as function argument instead of member
Changing the value of DateFormatPattern objects is not allowed, because
the same objects are used to analyze, parse and format a date. When the
a pattern can accept multiple terms (e.g. 'yyy', 'yy' and 'y'), there
might be a different accepted term during formatting and analyzing (e.g.
pattern is 'dd.MM.yy' but user enters '13.04.201').

Change-Id: I38e24410b31bdff89e65f67720d78d0ed8fd6da4
Signed-off-by: Thomas Plüss <thomas.pluess@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/95142
Tested-by: Hudson CI