iOS: Fix startup error when running on Tomcat
If the app is deployed on a tomcat and called from a link on the iOS
home screen, the app won't start. Instead "Initialization failed" is
shown. 
This happens because in that mode a persistent session cookie is
created. To make the name of the cookie configurable, the name is read
from the web.xml using SessionCookieConfig.getName(). If there is no
name set, the default name "JSESSIONID" should be used. When running
with Jetty it works fine because Jetty returns that default. Tomcat on
the other hand returns null which results in an error.

This only happens when running the app with the home screen mode. If
called using Safari it works fine because no persistent session cookie
will be created.