DatePicker: replace swipe and remove jquery mobile
The only thing needed by jquery mobile was the swipe event used by the
date picker. To make it obsolete the swipe handling for the date picker
has been reworked.

The new swipe handling is actually more like a drag swipe, which means
the scroll pane is moved while dragging. When the finger is released the
pane scrolls to the next month.

Also done: 
- Popup is now as height as its content.
- When changing the selection the date picker is not redrawn every time
anymore
- When holding and pressing the arrow down key the transition to the new
month is now smoother