Planner: activity is not selectable anymore when resource is updated
If the property activitySelectable is set to true, the activity may be
selected by clicking it. This is not possible anymore if an
updateResources event is sent in order to update a resource or activity.

Reason: link from the activity to the DOM element is not established on
an update event.

207929