Add AjaxCall.js to improve reconnect behavior for bad connections
To improve the user experience when the connection to the UI server is
bad (i.e. single requests fail sporadically), failed requests should be
tried again before failing ultimately and putting the session in
"offline" mode.

AjaxCall.js:
Provides a promise that is resolved when the ajax call has succeeded.
The promise is resolved if the call failed even after multiple retries.
After how many retries the call eventually succeeded (or ultimately
failed) is completely transparent to the caller.
This class also provides the static method isOfflineError() that was
previously inside Session.js.

Call.js:
Abstract base class for AjaxCall (could be used for other types of
"calls" that want to leverage the retry-before-fail feature).

RequestHistory.java:
To keep track of requests that were already processed, the request
counter was replaced by a history object that supports tracking
"missing" requests. This is necessary because the poll request now also
have a request sequence no. and might be "overtaken" by shorter user
requests.

177772

Change-Id: I2ae18984e8e36e138b1b22a5c81ab5b4cd0cabfa