Fixed problem with tooltips on detached widgets.
The problem occurred for instance on detached tables, on which a table
status has been set. This caused an UI error. The reason is that we
cannot determine the entryPoint() while the parent of the tooltip is
detached. The bugfix defers the render operation until the parent is
attached again. It also covers an earlier bugfix where the $anchor of
the tooltip was detached. The new solution does not require the
DetachHelper, so the expectations in the DetachHelperSpec have also been
changed.

207828