PopupWithHead: fix maxSizeAroundAnchor calculation
Use case: on the top right corner there is a menu. The menu is
positioned right at the border and there is only a small gap. There is
no logo as well (which is the case in the mobile mode). If the menu has
a sub menu, the animation was wrong and the sub menu items had an
ellipsis.
Reason: the function _calcMaxSizeAroundAnchor did not return the correct
values if the opening direction was 'left' instead of 'right'.

Also: If the PopupWithHead does not have a head, the default implementation of _calcMaxSizeAroundAnchor should be used. This makes the implementation of PopupWithHeadLayout simpler.