Outline: Explicitly set text color
Child pages inherit all properties of the summary cell in a table page,
including the "cssClass" property. While a colored text may look good in
the table, the same color is not appropriate in the outline, because the
desktop navigation background color is usually different from the table
background (e.g. dark vs. light).

Example: inactive values may be marked with a CSS class that displays
the text in gray color. If a drill-down is executed on such a value, the
corresponding tree node would be gray as well -> very bad contrast on
dark navigation background color.

My explicitly setting the text color in .outline > .tree-data >
.tree-node, accidental inheritance of styles from the table summary cell
is prevented. The CSS rules have to be formulated explicitly, if this
behavior should be overridden.

207129

Change-Id: I9d0cb031213ae72bb88a6428d2c9d601a246ac9f
Reviewed-on: https://git.eclipse.org/r/96108
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>