SplitBox: Improve splitter position
The middle of the splitter (or temporary splitter indicator) should
always be directly at the mouse cursor position. For best results, the
splitter width should be an odd number. The middle of the splitter is
1px, with an additional padding on both sides.

Note: The x splitter has a smaller padding at the right side, because
the right field will add a margin for the mandatory indicator by itself.

The splitter's CSS was changed to use a transparent background (replace
borders with background color by margins, the gray border is drawn
using a ::before element).

Change-Id: I629bb0a52d119a2eb60d8572de165e13329531fc