ModelAdapter: remove obsolete code
It seems that findModelAdapter is not necessary anymore. Instead
creating a local widget which has child widgets will fail in a remote
app.

Use case:
scout.create('GroupBox', {
  parent: scout.sessions[0].desktop,
  label: 'My Group Box',
  fields: [{
    objectType: 'StringField',
    label: 'My String Field'
  }]
});

This will fail because it tries to call getOrCreateAdapter with the
child model instead of a string.

Removing the findModelAdapter function fixes it.