Remove $parent from _render and make $parent optional
Remove $parent from the _render function because this.$parent is
available for every widget. There is no need to have a parameter $parent
which points to the same variable.

Make $parent optional when calling widget.render(). The $parent may be
resolved using this.parent. No need to always write
widget.render(this.$container).