ValueField: be more lenient when setting a value
If a value is set using setValue (numberfield or datefield), it should
not fail if the type is wrong. Instead, it should try to cast it and if
it is still the wrong type it may fail.

Simple reason: Typical case is to read a value out of a string field and
set it into a number field, or vice versa. The developer would have to
cast by himself which is annoying.

Also don't trigger a property change event if the value has not changed.

Necessary if two fields replicate its value, it would result in an
endless loop.