Tree: Ensure animation-wrapper is removed after batch insertion
When inserting batches of nodes, an $animationWrapper is created and
stored on the "insertBatch" object. The animation is started in the
insertBatchInVisibleNodes() method. However, in some cases, the nodes
are not visible and not animation is started. In this case, the
"animation complete" function must be called immediately, otherwise the
useless animation wrapper will stay in the DOM forever.

Change-Id: I770cd5dd1e0e598c60857f8f3cd666598db3ac55