ClipboardField: Fix detection of pasted bitmaps from the clipboard
When pasting bitmap data from the clipboard (not an image file), Chrome
and Firefox generate a dummy filename such as "image.png". To give the
model code a change to distinguish between bitmap data and image files,
a workaround has been implemented to set the image filename to ''.

The behavior was changed with Chrome 58:
https://bugs.chromium.org/p/chromium/issues/detail?id=361145
(It is unknown for how long Firefox already does the same.)

208956

Change-Id: Ice8a523ba84fbbf53fef674ccfc7e0545a6c326c