BeanField: Prevent unwanted change of value via 'format -> parse'
Under some conditions, it was possible to clear the contents of a bean
field by the user, even though the bean field is not actually editable.

The problem was caused because the value (a bean) was also used as a
displayText, but _readDisplayText() returned the empty string. Because
the displayText seems to have changed, acceptInput() would call
parseAndSetValue() and therefore overwrite the existing value.

Added Jasmine test for BeanField.

Change-Id: I5b3c84779859309a1fb9ce507735bc8eac0fce12