FocusContext: Prevent JS error when new focus elmnt cannot be determined
FocusContext tries to find the new focused element to scroll it into
view. Under some conditions, the new element cannot be determined.

Example in IE: mark text in a bean field, press Shift-TAB.
(For some reason, the bean field is the "activeElement" in Internet
Explorer, but not in other browsers.)

Formatted code.

Change-Id: I46e873199c079891efd7cfad29c646186e23becf