Popup opens in wrong direction with IE and Firefox sometimes
If the window is too small so that the poupup may not be fully
displayed, it considers switching the direction. If a menu of the main
menu bar is opened with IE and Firefox, it is always switched if it may
not be fully displayed. But the switch should only happen if there is
more space on the other side.

Reason is: IE and Firefox return a fractional value for the y position
of the menu. If that happens, the height of the menu will be fractional
as well, but if the height is read by popup.overlap() it will be rounded
up. This means Popup.overlap() returns 0.5 and the popup direction will
be switched.

208290