Menu: make IE trigger the :active state if the text or icon is pressed
If the text on a menu button is clicked, the active state is not
triggered (only IE). It works if the text of a regular button is
clicked.

The difference: menu buttons use a <div>, regular button a <button>
element.

Setting pointer-events to none fixes it, but only if display style is
not 'inline'.

207478