KeyStrokes: desktop keystrokes don't work anymore sometimes
Desktop keystrokes stop working if they were pressed while a modal
dialog was open. After that, even if the modal dialog had been closed,
the keystrokes did not work anymore.

Reason: KeyStrokeManager created a new keyStrokeContext object with
keystrokes which are available while a modal dialog is open. This new
keyStrokeContext replaced the closure variable and from now on the
context was broken.

Also made sure the new keyStrokeContext is of type KeyStrokeContext and
not just object (even though this had no effect on the bug).

209160