Desktop: move desktop keystroke registration from bench to desktop
There is no reason why the bench should register the keystrokes of the
desktop. This also enables dynamic key stroke changes, if the desktop
keystrokes change on the fly the are now properly updated.

Also set invokeAcceptInputOnActiveValueField to true on desktop
keystroke context because you never know what model keystrokes are
registered and if they need this behavior. And it was true on the bench.