NavigateDownButton: ensure updateEnabled is called when linked
The commit 24c363ac1f31709bd0a5bed9c73185cf6363c988 introduced the
additional condition to enable the button. This did not work when the
browser page was refreshed and then the table reload button clicked.
Reason: PageInit event came in the wrong order. 

To make it independent of the order a new event linkPageWithRow has been
introduced.