DatePicker: don't loose selected state when view date changes
If a date is selected and the view date changes (using scroll wheel or
the month switch buttons), the selected date was set to undefined. When
switching back to the original month, the previously selected day was
not visible anymore. Also, pressing DOWN or UP was not possible anymore
as well.