DateField: text does not match selection in picker sometimes
1. Focus empty date field, press down, delete text, press down again ->
The selected day has to be the current day, but the field is empty
because the selection in the picker was not removed.

2. Focus date field with a value, enter another date using keyboard
input (don't use arrow up/down), click the day which was selected when
the picker opened -> The newly entered input is written in the field
instead of the date selected by the mouse.