Fix datefield and numberfield error validation in sequence box
If a value was entered in the second field which is smaller than the one
in the first field, an error occurred, which is correct. If this was
done a second time the error vanished because the ui deleted the error
status but the server did not send it anew. Reason: The error status did
not change so the server had no reason to send it again.