Ensure that service calls don't terminate prematurely.
In commit 3899cc6f19 the network communication was switched to use the
Google HttpClient (backed by Apache HttpClient).
Because this implementation sets the default read timeout to 20 seconds,
longer running service calls are aborted and result in a
SocketTimedOutException.

To avoid this behaviour change, set the read timeout to infinity (0).

Change-Id: I7085fc33d1ae6a121181028101d2ffba9d036f20
Signed-off-by: Patrick Bänziger <patrick.baenziger@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/98304
Tested-by: Hudson CI
Reviewed-by: Matthias Otterbach <matthias.otterbach@bsi-software.com>