Correctly visualize "enabledComputed" in form fields
FormField.js:
All form fields have a second enabled-related property "enabledComputed"
that is calculated using the fields _and_ the parent's enabled flag.
When rendering the enabled state, this combined property should be used
on the $container as well (not only on the $field). Otherwise, fields
inside a disabled parent could _appear_ enabled, despite behaving as
disabled otherwise.

FileChooserField.js:
The "enabled" property of the inner FileInput widget must be
synchronized with the field's "enabledComputed" property, not "enabled".
Otherwise, the field would look disabled, but could be focused and
clicked.

Change-Id: I989af7fc2e52e948857083fef870b7c34fb24843