Add possibility to prevent double clicks on buttons and menus
When a user performs two consecutive clicks on a button or a menu within
a very short period of time, the action is executed twice. In some
cases, this is an unwanted behavior. A new property "preventDoubleClick"
was added to IButton and IMenu to disable it. The default value is
false. This default was chosen to support fast "add" or "trigger"
actions (e.g. increase a counter or insert many new rows in a table).

In older Scout versions, two flags (m_enabledProcessing and
m_enabledProcessingAction, respectively) were use to accomplish a
similar objective. Whenever the click/action handler was started, the
flag was set to false, and set back to true at the end of the execution.
A second call to the method had no effect. However, this logic no
longer works since the introduction of the HTML UI. The new UI strictly
allows only one UI event to work on the client model simultaneously
(subsequent click events are hold back by the UiSession lock).
Consequently, the "enabledProcessing" flag is never false, and no clicks
are prevented. Because the construct no works, the corresponding
properties and methods were removed, namely:
- AbstractButton: m_enabledProcessing
- IButton/AbstractButton: isEnabledProcessing()
- IButton/AbstractButton: setEnabledProcessing()
- IButton/AbstractButton: disarm()
- ButtonEvent: TYPE_DISARM
- AbstractAction: 
- IAction/AbstractAction: isEnabledProcessingAction

The following constructs are new:
- AbstractButton: getConfiguredPreventDoubleClick() [default: false]
- IButton: PROP_PREVENT_DOUBLE_CLICK
- IButton/AbstractButton: isPreventDoubleClick()
- IButton/AbstractButton: setPreventDoubleClick()
- AbstractMenu: getConfiguredPreventDoubleClick() [default: false]
- IMenu: PROP_PREVENT_DOUBLE_CLICK
- IMenu/AbstractMenu: isPreventDoubleClick()
- IMenu/AbstractMenu: setPreventDoubleClick()

209813

Change-Id: Ia461cb6eff85e85dd12030cb6ca36922f2a9f03e