Rename gridColumnCountHint to gridColumnCount
No need to use the suffix hint. This makes it more consistent with the
Scout JS layer.

All field grids now implement the interface ICompositeFieldGrid. All
composite fields now have a getter called getFieldGrid().

Warning: The previous function getGridColumnCount() returned the calculated column count of the grid, now it returns the configured one which is not necessarily the same. Example: the default of gridColumnCount is -1 which will typically result in 2 columns.