Ensure that KeyStrokes always reset themselves after deattaching.
Previously, it was the case that navigating away from a page with the
help of a KeyStroke via the use of keyboard buttons (e.g. (ENTER,
'keydown')) would cause that KeyStroke to unregister itself as a
listener before it received the matching (ENTER, 'keyup') event. When
the KeyStroke was reactivated, it was still incorrectly waiting for the
'keyup' instead of the 'keydown' event. The KeyStroke now correctly
resets itself on deattaching as well as on the 'keyup' event.
Also add tests, fix old tests. Some tests might still be broken using
PhantomJS, and have to be run on CI first to verify this fix.
209496

Change-Id: I80121428556084c11a649801ff1609a88abab488
Signed-off-by: Rudolf Schreier <Rudolf.Schreier@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/99460
Tested-by: Hudson CI
Reviewed-by: Andre Wegmueller <awe@bsiag.com>