jquery-scout: isVisible() should only consider 'display: none'
$.fn.isVisible() internally relies on the elemVisible() function. This
function was changed, so it only returns false when the CSS property
'display' has the value 'none', true otherwise. Otherwise, elements that
are drawn on screen but at not currently visible to the user (e.g.
because the container field is has loading=true, which sets 'visibility:
hidden' on all child elements) would be considered hidden. Layout code
could then not get the correct size of the element and the pixel-based
layout would be wrong.

The "scout.graphics.getVisibleSize()" method is not necessary anymore,
because scout.graphics.size() already returns (0,0), when a component is
not visible.

Change-Id: I024c83ff096031bb88588cdd8289c7e7b806648d