Mobile: prevent flickering when detail form is shown
At the time a node is selected it is not known, whether there is a
detail form or not because the form is created lazily. 
At the moment, the table row detail is created and shown, but if there
is a detail form it will be replaced as soon as the response returns.
The same applies for the menus: the response may contain additional
menus (belonging to the form).

To prevent flickering, the table row form and menus are only shown after
the response arrives.