Widget: don't destroy if removal is pending
Use case: smart field or tooltip is open on a search form in mobile
mode. The search form will be removed by a CSS animation. The user
clicks on the top of the screen so that the search form popup closes.
The smart field popup or tooltip stayed open because destroy() did not
remove it (due to isRemovalPending = true). It was also not removed
after the search form had been closed, because the parent / child link
was already removed by destroy().

-> Make sure parent child link is kept until the widget has really been
removed. Also make sure popups and tooltips are removed before the
animation starts, otherwise tooltip would still be visible for a short
time after the form had been closed.