Mobile: context menu popup sometimes not shown in outline
1. Show table row detail
2. Scroll down a little
3. Open a menu with submenus

Depending on the scroll position, the menu was not shown.

Reason: opening the menu added the class focused to the outline. The
class removed the border of the ::before element of the parent of the
selected node. This changed the size a little and made the data scroll
which closed the popup right after it was opened.

210281