SmartField2: fixed Selenium SmartField2Test
Various issues solved: most notable was a bug where the wrong result was
used when a LookupCall was done. Now we check that only the latest
result is used (see RemoteLookupRequest.js)

The TreeProposalChooser2 selected the first node in the tree, which is
wrong when we search for text (compare with original SmartField). No
node should be selected, because a hierarchical lookup always returns
all parent rows for the node we're looking for.

ContentAssistSearchParam: store original search text as the user has 
typed in the UI. The new property searchQuery has the value of the old
searchText property, where an empty searchText is replaced with the
wildcard character.