Don't show tooltip if text is not truncated
Tooltips are sometimes shown even if the text is not truncated,
especially on table header cells and checkbox labels.

Reasons:
1. getBoundingClientRect includes border, clientWidth does not.
Comparing these values will always return true if the element has a
border (as the table header cell has).

2. If the element has a max-width, resetting the width has no effect ->
reset max-width as well (relevant for checkbox labels).

3. The IE sometimes returns strange scrollWidth values. One example: it
returned scrollWidth=49px and clientWidth=48px for the label 'Default',
but did not show an ellipsis. At first I thought it is because the width
is actually fractional (getBoundingClientRect().width returns 48.2), but
other label have a fractional width as well but have correct
scrollWidth. To fix it, scrollWidth is not used anymore with IE.